+++ src/blcmm/utilities/Options.java
@@ -33,6 +33,7 @@ import blcmm.gui.theme.Theme;
 import blcmm.gui.theme.ThemeManager;
 import blcmm.model.PatchType;
 import blcmm.utilities.options.*;
+import java.awt.event.MouseEvent;
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.nio.file.Paths;
@@ -88,6 +89,200 @@ public class Options extends OptionsBase {
     }
 
     /**
+     * An enum describing the actions which can be taken when processing a
+     * mouse click on an object link (in an edit panel or OE panel).  The
+     * valid options are: No action, Open in Current Tab, or Open in New Tab.
+     *
+     * The implemented interface list here is a bit wonky; I can't help but
+     * thinking I'm way overcomplicating this feature.
+     */
+    public enum MouseLinkAction implements OptionEnum, SelectionOptionData {
+        None("No Action"),
+        Current("Current Tab"),
+        New("New Tab");
+
+        private final String description;
+
+        private MouseLinkAction(String description) {
+            this.description = description;
+        }
+
+        @Override
+        public String getDescription() {
+            return this.description;
+        }
+
+        @Override
+        public String toSaveString() {
+            return this.name();
+        }
+
+        @Override
+        public MouseLinkAction getRawData() {
+            return this;
+        }
+
+    }
+
+    /**
+     * A helper class to handle configuration of mouse object-name link
+     * behaviors, since we want to configure/handle more than one mouse click
+     * configuration.  This sets up a block of entries on our settings screen,
+     * and also handles dealing with MouseEvents.
+     */
+    public class MouseLinkSection {
+
+        private final Options options;
+        private final String identifier;
+
+        /**
+         * A new section to configure mouse-object-name-link behavior.
+         *
+         * @param options The Options object we live in
+         * @param section The section in which to display the options
+         * @param identifier An internal identifier for the button; will show
+         *      up in the setting names in our settings file.
+         * @param header The header to print in front of the section
+         * @param defaultButton The default button for this config
+         * @param defaultNumClicks The default number of clicks to trigger on
+         * @param defaultBase The default action for an unmodified click
+         * @param defaultCtrl The action to take when the user does a Ctrl-click
+         * @param defaultMeta The action to take when the user does a Meta-click
+         * @param defaultShift The action to take when the user does a Shift-click
+         * @param defaultAlt The action to take when the user does a Alt-click
+         */
+        public MouseLinkSection(Options options,
+                Option.Shown section,
+                String identifier,
+                String header,
+                int defaultButton,
+                int defaultNumClicks,
+                MouseLinkAction defaultBase,
+                MouseLinkAction defaultCtrl,
+                MouseLinkAction defaultMeta,
+                MouseLinkAction defaultShift,
+                MouseLinkAction defaultAlt
+                ) {
+            this.options = options;
+            this.identifier = identifier;
+
+            options.registerOption(new SectionHeaderOption(options, section, header));
+
+            options.registerOption(new IntOption(
+                    options,
+                    "inputMouseLink" + identifier + "Button", defaultButton,
+                    section, "Mouse Button",
+                    1, 12,
+                    null,
+                    "Mouse button for clicking on object links"
+            ));
+
+            options.registerOption(new IntOption(
+                    options,
+                    "inputMouseLink" + identifier + "Clicks", defaultNumClicks,
+                    section, "      Number of Clicks",
+                    0, 4,
+                    null,
+                    "Number of clicks to require to load an object link with this button"
+            ));
+
+            options.registerOption(SelectionOption.createEnumSelectionOption(options,
+                    "inputMouseLink" + identifier + "Base", defaultBase,
+                    section,
+                    "      Base Action",
+                    null,
+                    "Base action to take when clicked without any keyboard modifiers",
+                    MouseLinkAction.values()
+            ));
+
+            options.registerOption(SelectionOption.createEnumSelectionOption(options,
+                    "inputMouseLink" + identifier + "Ctrl", defaultCtrl,
+                    section,
+                    "        ...with Ctrl",
+                    null,
+                    "Action to take when clicked with Ctrl held down",
+                    MouseLinkAction.values()
+            ));
+
+            options.registerOption(SelectionOption.createEnumSelectionOption(options,
+                    "inputMouseLink" + identifier + "Meta", defaultMeta,
+                    section,
+                    "        ...with Meta/⌘",
+                    null,
+                    "Action to take when clicked with Meta/⌘     held down",
+                    MouseLinkAction.values()
+            ));
+
+            options.registerOption(SelectionOption.createEnumSelectionOption(options,
+                    "inputMouseLink" + identifier + "Shift", defaultShift,
+                    section,
+                    "        ...with Shift",
+                    null,
+                    "Action to take when clicked with Shift held down",
+                    MouseLinkAction.values()
+            ));
+
+            options.registerOption(SelectionOption.createEnumSelectionOption(options,
+                    "inputMouseLink" + identifier + "Alt", defaultAlt,
+                    section,
+                    "        ...with Alt",
+                    null,
+                    "Action to take when clicked with Alt held down",
+                    MouseLinkAction.values()
+            ));
+
+        }
+
+        /**
+         * Checks the given MouseEvent to see if it matches our initial
+         * requirements (ie: mouse button and click count)
+         *
+         * @param e The event to check
+         * @return True if we should handle this event, false otherwise
+         */
+        private boolean matches(MouseEvent e) {
+            return (e.getButton() == this.options.getIntOptionData("inputMouseLink" + this.identifier + "Button")
+                    && e.getClickCount() == this.options.getIntOptionData("inputMouseLink" + this.identifier + "Clicks")
+                    );
+        }
+
+        /**
+         * Gets the appropriate action to take for the specified MouseEvent, or
+         * null if we shouldn't handle the event.
+         *
+         * @param e The event to check
+         * @return A MouseLinkAction to take, or null.
+         */
+        public MouseLinkAction getAction(MouseEvent e) {
+            // The data-retrieval syntax here is absurd.  There's gotta be a much
+            // cleaner way of doing this.  I suspect that my whole OptionEnum
+            // stuff is way overthinking things.
+            if (this.matches(e)) {
+                if (e.isControlDown()) {
+                    return (MouseLinkAction)((SelectionOptionData)this.options.getSelectionOptionData("inputMouseLink" + this.identifier + "Ctrl", MouseLinkAction.class)).getRawData();
+                } else if (e.isMetaDown()) {
+                    return (MouseLinkAction)((SelectionOptionData)this.options.getSelectionOptionData("inputMouseLink" + this.identifier + "Meta", MouseLinkAction.class)).getRawData();
+                } else if (e.isShiftDown()) {
+                    return (MouseLinkAction)((SelectionOptionData)this.options.getSelectionOptionData("inputMouseLink" + this.identifier + "Shift", MouseLinkAction.class)).getRawData();
+                } else if (e.isAltDown()) {
+                    return (MouseLinkAction)((SelectionOptionData)this.options.getSelectionOptionData("inputMouseLink" + this.identifier + "Alt", MouseLinkAction.class)).getRawData();
+                } else {
+                    return (MouseLinkAction)((SelectionOptionData)this.options.getSelectionOptionData("inputMouseLink" + this.identifier + "Base", MouseLinkAction.class)).getRawData();
+                }
+            } else {
+                return null;
+            }
+        }
+
+    }
+
+    /**
+     * An array of MouseLinkSections which we'll use to check MouseEvents for,
+     * when the user clicks on object name links.
+     */
+    private final MouseLinkSection[] mouseLinkSections;
+
+    /**
      * HashSet which collects the currently-active OE Search/Refs categories.
      */
     private HashSet<OESearch> activeSearchCategories = new HashSet<>();
@@ -125,10 +320,11 @@ public class Options extends OptionsBase {
                 Option.Shown.SETTINGS, "Theme",
                 "setTheme", "Change " + Meta.NAME + "'s color theme",
                 "checkThemeSwitchAllowed",
-                ThemeManager.getAllInstalledThemes().toArray(new Theme[0]), s -> {
-            Theme t = ThemeManager.getTheme(s);
-            return t == null ? ThemeManager.getTheme("dark") : t;
-        }
+                ThemeManager.getAllInstalledThemes().toArray(new Theme[0]),
+                s -> {
+                    Theme t = ThemeManager.getTheme(s);
+                    return t == null ? ThemeManager.getTheme("dark") : t;
+                }
         ));
 
         this.registerOption(new BooleanOption(
@@ -197,6 +393,37 @@ public class Options extends OptionsBase {
                 + " with the 'full' object name, including class type.<br/>"
                 + "<b>Note:</b> This will limit autocomplete results to the specified type,"
                 + " when the class type is present."));
+
+        // Now options in the Input Settings area
+
+        this.mouseLinkSections = new MouseLinkSection[] {
+            new MouseLinkSection(this,
+                    Option.Shown.INPUT,
+                    "Primary",
+                    "Primary Button",
+                    1,
+                    2,
+                    MouseLinkAction.Current,
+                    MouseLinkAction.New,
+                    MouseLinkAction.New,
+                    MouseLinkAction.Current,
+                    MouseLinkAction.Current
+                    ),
+            new MouseLinkSection(this,
+                    Option.Shown.INPUT,
+                    "Secondary",
+                    "Secondary Button",
+                    2,
+                    2,
+                    MouseLinkAction.New,
+                    MouseLinkAction.New,
+                    MouseLinkAction.New,
+                    MouseLinkAction.New,
+                    MouseLinkAction.New
+                    ),
+        };
+
+        // Now options in the OE Data area
 
         this.registerOption(new BooleanOption(this, OptionNames.oeSearchActions.toString(), true,
                 Option.Shown.OE, "Actions Data",
@@ -274,6 +501,8 @@ public class Options extends OptionsBase {
                 "updateOESearchCategories",
                 "Search \"WillowData\" classes during fulltext and refs searches.  "
                 + "This package is only useful in specific circumstances."));
+
+        // Now Dangerous Settings
 
         this.registerOption(new BooleanOption(this, OptionNames.saveAsOffline.toString(), true,
                 Option.Shown.DANGEROUS, "Save patch files in 'Offline' Mode",
@@ -729,5 +958,22 @@ public class Options extends OptionsBase {
 
     public Set<OESearch> getOESearchCategories() {
         return this.activeSearchCategories;
+    }
+
+    /**
+     * Given a MouseEvent, return a MouseLinkAction if we have a mouse-link
+     * handler which should handle the event, or null otherwise.
+     *
+     * @param e The event to check
+     * @return a MouseLinkAction, or null
+     */
+    public MouseLinkAction processMouseLinkClick(MouseEvent e) {
+        for (MouseLinkSection mls : this.mouseLinkSections) {
+            MouseLinkAction a = mls.getAction(e);
+            if (a != null) {
+                return a;
+            }
+        }
+        return null;
     }
 }